figured things out
